pyproject.toml
[project]
name = "invoice-processor"
version = "0.1.0"
description = "AI-powered invoice processing and organization system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.5.3",
    "psutil>=7.0.0",
    "pydantic>=2.11.7",
    "pydantic-ai[ollama]>=0.7.5",
    "pypdf>=6.0.0",
    "pyyaml>=6.0.2",
    "textual>=5.3.0",
]

[project.scripts]
invoice-processor = "src.invoice_processor.cli:main"

[project.optional-dependencies]
gpu = [
    "GPUtil>=1.4.0",
    "pynvml>=11.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/invoice_processor"]
